requirements.txt
flask==3.1.*
flask-sqlalchemy==3.1.*
flask-migrate==4.1.*
flask-cors==5.0.*
psycopg2-binary==2.9.*
pyjwt[crypto]==2.10.*
python-dotenv==1.1.*
beautifulsoup4==4.13.*
requests==2.32.*
gunicorn==23.0.*
